[type:refactor] refactor shenyu-admin sync data listener.
